[add] step9 return文

+++ src/9cc.h
@@ -74,7 +74,7 @@ struct Node
     Node *next;    // 次のノード
     Node *lhs;     // 左辺
     Node *rhs;     // 右辺
-    int val;       // kindがND_NUMの場合のみ使う
+    long val;      // kindがND_NUMの場合のみ使う
     int offset;    // kindがND_LVARの場合のみ使う。ローカル変数のベースポインタからのオフセット
 };
